Одежда из дополнений перестала заворачиваться в байт
Номер вещи сквозной по всем дополнениям сразу: они идут следом за основными, и у
торса их нынче за четыре сотни. Умещённый в байт, номер молча заворачивался —
просивший триста получал сорок четвёртое, — и вся одежда из дополнений была
недостижима при том, что своих нативов ей не нужно никаких.

Живая игра подтвердила обе стороны: сервер просит трёхсотую, игра отвечает
трёхсотой.

`setDlcClothes`, `setDlcProp` и `getDlcClothes` при этом говорят о себе вслух и
отсылают к `setClothes`: пары «хеш набора и место в нём» игра обратно не отдаёт,
нативов `..._COLLECTION_...` нет ни в открытой базе CitizenFX, ни в таблице
alt:V. Вопрос отказывает, распоряжения предупреждают — по тому же разделению,
что и весь слой.

+++ client/src/game/ped_appearance.cpp
@@ -68,8 +68,8 @@ shared::PlayerAppearance PedAppearance::read(int ped) const {
         const auto component = static_cast<int>(slot);
 
         appearance.components[slot] = shared::PedComponent{
-            .drawable = static_cast<std::uint8_t>(invokeNative<int>(getDrawable_, ped, component)),
-            .texture = static_cast<std::uint8_t>(invokeNative<int>(getTexture_, ped, component)),
+            .drawable = static_cast<std::uint16_t>(invokeNative<int>(getDrawable_, ped, component)),
+            .texture = static_cast<std::uint16_t>(invokeNative<int>(getTexture_, ped, component)),
             .palette = static_cast<std::uint8_t>(invokeNative<int>(getPalette_, ped, component)),
         };
     }
@@ -80,8 +80,8 @@ shared::PlayerAppearance PedAppearance::read(int ped) const {
         // Игра отвечает минус единицей, когда в слоте пусто, и это же значение
         // мы передаём дальше: оно означает «снять», а не «надеть нулевую вещь».
         appearance.props[slot] = shared::PedProp{
-            .drawable = static_cast<std::int8_t>(invokeNative<int>(getProp_, ped, prop)),
-            .texture = static_cast<std::int8_t>(invokeNative<int>(getPropTexture_, ped, prop)),
+            .drawable = static_cast<std::int16_t>(invokeNative<int>(getProp_, ped, prop)),
+            .texture = static_cast<std::int16_t>(invokeNative<int>(getPropTexture_, ped, prop)),
         };
     }
